requirements.txt
configparser~=5.0.2
elasticsearch~=7.11.0
imutils~=0.5.4
jsonschema~=3.2.0
konlpy~=0.5.2
matplotlib~=3.3.4
mysql-connector-python~=8.0.25
numpy~=1.19.2
opencv-python~=4.5.2.54
pandas~=1.1.5
pathlib~=1.0.1
Pillow~=8.2.0
PyKomoran~=0.1.5
PyMySQL~=0.10.1
pytagcloud~=0.3.5
PyYAML~=5.4.1
requests~=2.25.1
scikit-image~=0.18.1
scikit-learn~=0.24.2
scipy~=1.6.2
simplejson~=3.17.2
six~=1.15.0
sklearn~=0.0
tensorflow-gpu~=2.5.0
torch~=1.9.0
torchvision~=0.10.0
tqdm~=4.49.0
utils~=1.0.1
wandb~=0.10.30
Werkzeug~=1.0.1
pygame
argparse~=1.4.0
